// types/RequestNavigation.ts
export interface RequisicaoParams {
    SolicitacaoId?: string;
    LocadorId?: string;
    LocatarioId?: string;
    CarroID?: string;
    SelectedModalidade?: string;
    DataInicio?: string;
    DataTermino?: string;
    TotalValor?: string,
    TotalDias?: string;
    Dia?: string;
    Estado?: string;
    Descricao?: string;
}